Sooooo, i've got _an_ algorithm, honestly it's pretty good (apart for the obvious non finished things like don't handle buzzing_grid.json) it generates intersting levels. There is a thing altough it can generate things like:
    o-o
    |   |
o---o
    |
   o

where there is a little cross over, i don't really know if it's a bad thing as the transistion between section are discontinuous so it's rp to allow cross over... Anyway i don't think it would be that bad to fix so it's good.

With few tweeks i can add specials rooms for gameplay purposes so that's good too.
I Might want to revisit my blue noise generator(for now it's poisson disk distribution) it's quite slow...
And my minimum spanning tree wan be optimise too...

In a few sections i didn't set correctly the portals drop_pos so you end up in the void...